refactor: minor refactoring and route fixes
