+++ src/shared/libs/file-reader/file-reader.constant.ts
@@ -1,0 +1,1 @@
+export const CHUNK_SIZE = 16384; //16Kb
