+++ src/shared/libs/file-reader/tsv-file-reader.ts
@@ -1,9 +1,8 @@
 import {EventEmitter} from 'node:events';
 import {createReadStream} from 'node:fs';
 
+import {CHUNK_SIZE} from './file-reader.constant.js';
 import {FileReader} from './file-reader.interface.js';
-
-const CHUNK_SIZE = 16384; //16Kb
 
 export class TSVFileReader extends EventEmitter implements FileReader {
   constructor(
