+++ src/shared/libs/rest/middleware/middleware.interface.ts
@@ -1,5 +1,5 @@
 import {NextFunction, Request, Response} from 'express';
 
 export interface Middleware {
-  execute(req: Request, res: Response, next: NextFunction): void;
+  execute(req: Request, res: Response, next: NextFunction): Promise<void>;
 }
