+++ src/shared/modules/user/user.http
@@ -31,7 +31,7 @@ Content-Type: application/json
 ###
 
 ##Обновить информацию о пользователе
-PATCH http://localhost:6000/users/6542788fb2ebbcd60fa229ec HTTP/1.1
+PATCH http://localhost:6000/users/654aa5c3c40da95599f4a3c8 HTTP/1.1
 Content-Type: application/json
 
 {
@@ -56,7 +56,7 @@ Content-Type: image/png
 
 ##Проверить аутентификацию
 GET http://localhost:6000/users/login HTTP/1.1
-Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6ImJ1Z0Bkb250ZmVhdC5jb20iLCJuYW1lIjoiUUF0ZXN0RnJlYWsiLCJpZCI6IjY1NDI3ODhmYjJlYmJjZDYwZmEyMjllYyIsImlhdCI6MTY5ODg1NTA1OSwiZXhwIjoxNjk5MDI3ODU5fQ.8QnZLPtrI_HVCeWloDH11isybxnUN2k5kzL89f2IFIU
+Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6ImFub25AYW5vbmltb3VzLmNvbSIsIm5hbWUiOiJBbm9uVXNlciIsImlkIjoiNjU0YWE1YzNjNDBkYTk1NTk5ZjRhM2M4IiwiaWF0IjoxNjk5MzkwOTM0LCJleHAiOjE2OTk1NjM3MzR9.3VFVmxrHcjv3KP_Hp7A_aPD5pjhjA4qw0ZJ9D33SRgE
 
 ###
